+++ previous-versions/sync-for-commerce-version-1/src/sdk/models/shared/creditnotelineitem.ts
@@ -13,10 +13,16 @@ import { TrackingCategoryRefsitems } from "./trackingcategoryrefsitems";
 import { Expose, Type } from "class-transformer";
 
 export class CreditNoteLineItemTrackingProjectReference extends SpeakeasyBase {
+    /**
+     * Unique identifier to the project reference.
+     */
     @SpeakeasyMetadata()
     @Expose({ name: "id" })
     id: string;
 
+    /**
+     * The project's name.
+     */
     @SpeakeasyMetadata()
     @Expose({ name: "name" })
     name?: string;
@@ -36,10 +42,16 @@ export class CreditNoteLineItemTracking extends SpeakeasyBase {
     @Type(() => AccountingCustomerRef)
     customerRef?: AccountingCustomerRef;
 
+    /**
+     * Defines if the bill or bill credit note is billed/rebilled to a project.
+     */
     @SpeakeasyMetadata()
     @Expose({ name: "isBilledTo" })
     isBilledTo: BilledToType;
 
+    /**
+     * Defines if the bill or bill credit note is billed/rebilled to a project.
+     */
     @SpeakeasyMetadata()
     @Expose({ name: "isRebilledTo" })
     isRebilledTo: BilledToType;
@@ -92,6 +104,9 @@ export class CreditNoteLineItem extends SpeakeasyBase {
     @Expose({ name: "discountPercentage" })
     discountPercentage?: number;
 
+    /**
+     * The credit note is a direct income if `True`.
+     */
     @SpeakeasyMetadata()
     @Expose({ name: "isDirectIncome" })
     isDirectIncome?: boolean;