+++ previous-versions/sync-for-commerce-version-1/src/sdk/models/shared/datatypefeature.ts
@@ -3,56 +3,9 @@
  */
 
 import { SpeakeasyBase, SpeakeasyMetadata } from "../../../internal/utils";
+import { DataType } from "./datatype";
 import { SupportedFeature } from "./supportedfeature";
 import { Expose, Type } from "class-transformer";
-
-/**
- * Available Data types
- */
-export enum DataTypeFeatureDataTypes {
-    AccountTransactions = "accountTransactions",
-    BalanceSheet = "balanceSheet",
-    BankAccounts = "bankAccounts",
-    BankTransactions = "bankTransactions",
-    BillCreditNotes = "billCreditNotes",
-    BillPayments = "billPayments",
-    Bills = "bills",
-    CashFlowStatement = "cashFlowStatement",
-    ChartOfAccounts = "chartOfAccounts",
-    Company = "company",
-    CreditNotes = "creditNotes",
-    Customers = "customers",
-    DirectCosts = "directCosts",
-    DirectIncomes = "directIncomes",
-    Invoices = "invoices",
-    Items = "items",
-    JournalEntries = "journalEntries",
-    Journals = "journals",
-    PaymentMethods = "paymentMethods",
-    Payments = "payments",
-    ProfitAndLoss = "profitAndLoss",
-    PurchaseOrders = "purchaseOrders",
-    SalesOrders = "salesOrders",
-    Suppliers = "suppliers",
-    TaxRates = "taxRates",
-    TrackingCategories = "trackingCategories",
-    Transfers = "transfers",
-    BankingAccountBalances = "banking-accountBalances",
-    BankingAccounts = "banking-accounts",
-    BankingTransactionCategories = "banking-transactionCategories",
-    BankingTransactions = "banking-transactions",
-    CommerceCompanyInfo = "commerce-companyInfo",
-    CommerceCustomers = "commerce-customers",
-    CommerceDisputes = "commerce-disputes",
-    CommerceLocations = "commerce-locations",
-    CommerceOrders = "commerce-orders",
-    CommercePaymentMethods = "commerce-paymentMethods",
-    CommercePayments = "commerce-payments",
-    CommerceProductCategories = "commerce-productCategories",
-    CommerceProducts = "commerce-products",
-    CommerceTaxComponents = "commerce-taxComponents",
-    CommerceTransactions = "commerce-transactions",
-}
 
 /**
  * Describes support for a given datatype and associated operations
@@ -63,7 +16,7 @@ export class DataTypeFeature extends SpeakeasyBase {
      */
     @SpeakeasyMetadata()
     @Expose({ name: "dataType" })
-    dataType?: DataTypeFeatureDataTypes;
+    dataType?: DataType;
 
     @SpeakeasyMetadata({ elemType: SupportedFeature })
     @Expose({ name: "supportedFeatures" })