+++ previous-versions/sync-for-commerce-version-1/src/sdk/models/shared/productvariant.ts
@@ -108,6 +108,9 @@ export class ProductVariant extends SpeakeasyBase {
     @Expose({ name: "sourceModifiedDate" })
     sourceModifiedDate?: string;
 
+    /**
+     * The status of the product variant.
+     */
     @SpeakeasyMetadata()
     @Expose({ name: "status" })
     status?: ProductVariantStatus;