+++ previous-versions/sync-for-commerce-version-1/RELEASES.md
@@ -19,3 +19,13 @@ Based on:
 - [typescript v0.2.0] previous-versions/sync-for-commerce-version-1
 ### Releases
 - [NPM v0.2.0] https://www.npmjs.com/package/@codat/sync-for-commerce-version-1/v/0.2.0 - previous-versions/sync-for-commerce-version-1
+
+## 2023-10-17 13:37:11
+### Changes
+Based on:
+- OpenAPI Doc 1.1 https://raw.githubusercontent.com/codatio/oas/main/yaml/Codat-Sync-Commerce-v1.yaml
+- Speakeasy CLI 1.100.2 (2.159.2) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [typescript v0.3.0] previous-versions/sync-for-commerce-version-1
+### Releases
+- [NPM v0.3.0] https://www.npmjs.com/package/@codat/sync-for-commerce-version-1/v/0.3.0 - previous-versions/sync-for-commerce-version-1